orders/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.7
django-allauth==0.53.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
drf-spectacular==0.26.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
